学生模块后台接口全部完成，测试成功

+++ src/main/java/cn/wisdsoft/electivesystem/pojo/CurriculumDo.java
@@ -1,6 +1,7 @@
 package cn.wisdsoft.electivesystem.pojo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * @ Author     ：高伟萌.
@@ -48,7 +49,7 @@ public class CurriculumDo implements Serializable {
     /**
      * 学分
      */
-    private Double credit;
+    private BigDecimal credit;
 
     /**
      * 上课时间
@@ -123,11 +124,11 @@ public class CurriculumDo implements Serializable {
         this.classHour = classHour;
     }
 
-    public Double getCredit() {
+    public BigDecimal getCredit() {
         return credit;
     }
 
-    public void setCredit(Double credit) {
+    public void setCredit(BigDecimal credit) {
         this.credit = credit;
     }
